+++ otreeutils/surveys.py
@@ -35,12 +35,14 @@ def generate_likert_field(labels, widget=None):
     return partial(models.IntegerField, widget=widget, choices=choices)
 
 
-def generate_likert_table(likert_labels, questions, form_name=None, help_texts=None, widget=None, **kwargs):
+def generate_likert_table(labels, questions, form_name=None, help_texts=None, widget=None, use_likert_scale=True,
+                          make_label_tag=False, **kwargs):
     """
-    Generate a table with Likert scales between 1 and `len(likert_labels)` in each row for questions supplied with
+    Generate a table with Likert scales between 1 and `len(labels)` in each row for questions supplied with
     `questions` as list of tuples (field name, field label).
     Optionally provide `help_texts` which is a list of help texts for each question (hence must be of same length
     as `questions`.
+    If `make_label_tag` is True, then each label is surrounded by a <label>...</label> tag, otherwise it's not.
     Optionally set `widget` (default is `RadioSelect`).
     """
     if not help_texts:
@@ -52,17 +54,23 @@ def generate_likert_table(likert_labels, questions, form_name=None, help_texts=N
     if len(help_texts) != len(questions):
         raise ValueError('Number of questions must be equal to number of help texts.')
 
-    likert_field = generate_likert_field(likert_labels, widget=widget)
+    if use_likert_scale:
+        field_generator = generate_likert_field(labels, widget=widget)
+        header_labels = labels
+    else:
+        field_generator = partial(models.StringField, choices=labels, widget=widget or widgets.RadioSelectHorizontal)
+        header_labels = [t[1] for t in labels]
 
     fields = []
     for (field_name, field_label), help_text in zip(questions, help_texts):
         fields.append((field_name, {
             'help_text': help_text,
             'label': field_label,
-            'field': likert_field(),
+            'make_label_tag': make_label_tag,
+            'field': field_generator(),
         }))
 
-    form_def = {'form_name': form_name, 'fields': fields, 'render_type': 'table', 'header_labels': likert_labels}
+    form_def = {'form_name': form_name, 'fields': fields, 'render_type': 'table', 'header_labels': header_labels}
     form_def.update(dict(**kwargs))
 
     return form_def
@@ -141,6 +149,11 @@ class SurveyPage(ExtendedPage):
     field_labels = {}
     field_help_text = {}
     field_help_text_below = {}
+    field_make_label_tag = {}
+    field_input_prefix = {}
+    field_input_suffix = {}
+    field_widget_attrs = {}
+    field_condition_javascript = {}
     field_forms = {}
     forms_opts = {}
     form_label_suffix = ':'
@@ -151,11 +164,15 @@ class SurveyPage(ExtendedPage):
         survey_defs = player_cls.get_survey_definitions()[page_idx]
         cls.form_model = player_cls
         cls.page_title = survey_defs['page_title']
-        cls.form_label_suffix = survey_defs.get('form_label_suffix', ':')
+        cls.form_label_suffix = survey_defs.get('form_label_suffix', '')
 
         cls.field_labels = {}
         cls.field_help_text = {}
         cls.field_help_text_below = {}
+        cls.field_input_prefix = {}
+        cls.field_input_suffix = {}
+        cls.field_widget_attrs = {}
+        cls.field_condition_javascript = {}
         cls.field_forms = {}
         cls.forms_opts = {}
         cls.form_fields = []
@@ -164,6 +181,11 @@ class SurveyPage(ExtendedPage):
             cls_.field_labels[field_name] = qdef.get('text', qdef.get('label', ''))
             cls_.field_help_text[field_name] = qdef.get('help_text', '')
             cls_.field_help_text_below[field_name] = qdef.get('help_text_below', False)
+            cls_.field_make_label_tag[field_name] = qdef.get('make_label_tag', False)
+            cls_.field_input_prefix[field_name] = qdef.get('input_prefix', '')
+            cls_.field_input_suffix[field_name] = qdef.get('input_suffix', '')
+            cls_.field_widget_attrs[field_name] = qdef.get('widget_attrs', {})
+            cls_.field_condition_javascript[field_name] = qdef.get('condition_javascript', '')
             cls_.form_fields.append(field_name)
             cls_.field_forms[field_name] = form_name
 
@@ -205,7 +227,16 @@ class SurveyPage(ExtendedPage):
             form_name = self.field_forms[field_name]
 
             field.label = self.field_labels[field_name]
-            field.help_text = (self.field_help_text_below[field_name], self.field_help_text[field_name])
+            field.help_text = {  # abusing the help text attribute here for arbitrary field options
+                'help_text': self.field_help_text[field_name],
+                'help_text_below': self.field_help_text_below[field_name],
+                'make_label_tag': self.field_make_label_tag[field_name],
+                'input_prefix': self.field_input_prefix[field_name],
+                'input_suffix': self.field_input_suffix[field_name],
+                'condition_javascript': self.field_condition_javascript[field_name],
+            }
+
+            field.widget.attrs.update(self.field_widget_attrs[field_name])
 
             if form_name not in survey_forms:
                 survey_forms[form_name] = {'fields': [], 'form_opts': self.forms_opts.get(form_name, {})}